filter has appearance of search. having trouble integrating into UI
