+++ client/components/index.js
@@ -5,4 +5,6 @@
  */
 export {default as Main} from './main'
 export {default as UserHome} from './user-home'
+export {default as Search} from './Search'
+export {default as Queue} from './Queue'
 export {Login, Signup} from './auth-form'
