+++ package.json
@@ -47,6 +47,8 @@
     "react-flip-move": "^3.0.0",
     "react-redux": "^5.0.2",
     "react-router-dom": "^4.1.1",
+    "react-simpletabs": "^0.7.0",
+    "react-tabs": "^2.2.1",
     "redux": "^3.6.0",
     "redux-logger": "^2.8.1",
     "redux-thunk": "^2.2.0",
